refactor tile actions: javadocs, GoToJailAction test, checkstyle formatting
