+++ src/main/java/edu/ntnu/idi/idatt/domain/action/snakesandladders/ResetAction.java
@@ -6,35 +6,41 @@ import edu.ntnu.idi.idatt.domain.entity.Tile;
 import edu.ntnu.idi.idatt.utils.Validation;
 
 /**
- * ResetAction is a class that represents the action of moving the player to the start tile
- * when the player lands on a reset tile.
+ * Represents an action triggered when a {@link Player} lands on a reset tile.
+ * This action causes the player to return to their start tile.
+ *
+ * <br>
+ * <b>Note:</b> The player must have a defined start tile.
  *
  * @version 0.1
  * @since 0.1
  * @author Gilianne Reyes
  */
 public class ResetAction implements TileAction {
-    public static final String actionType = "ResetAction";
+  public static final String actionType = "ResetAction";
 
-    /**
-     * Moves the player to the start tile.
-     *
-     * @param player is the player that landed on the tile.
-     */
-    @Override
-    public void perform(Player player) {
-        Validation.validateNonNull(player, "Player");
-        Tile startTile = player.getStartTile();
-        player.placeOnTile(startTile);
-    }
+  /**
+   * Moves the specified player back to their start tile, simulating a reset.
+   *
+   * @param player is the player who landed on the reset tile.
+   *
+   * @throws IllegalArgumentException if the player is {@code null}.
+   * @throws IllegalStateException if the player's start tile is {@code null}.
+   */
+  @Override
+  public void perform(Player player) {
+    Validation.validateNonNull(player, "Player");
+    Tile startTile = player.getStartTile();
+    player.placeOnTile(startTile);
+  }
 
-    /**
-     * Retrieves the type of the action, which is "ResetAction".
-     *
-     * @return the type of the action.
-     */
-    @Override
-    public String getActionType() {
-        return actionType;
-    }
+  /**
+   * Retrieves a string identifier describing the {@link ResetAction}.
+   *
+   * @return a non-null string representing the action type.
+   */
+  @Override
+  public String getActionType() {
+    return actionType;
+  }
 }
