+++ src/main/java/edu/ntnu/idi/idatt/domain/action/snakesandladders/SkipTurnAction.java
@@ -1,46 +1,44 @@
-    package edu.ntnu.idi.idatt.domain.action.snakesandladders;
+package edu.ntnu.idi.idatt.domain.action.snakesandladders;
 
-    import edu.ntnu.idi.idatt.domain.action.TileAction;
-    import edu.ntnu.idi.idatt.domain.entity.Player;
-    import edu.ntnu.idi.idatt.utils.Validation;
+import edu.ntnu.idi.idatt.domain.action.TileAction;
+import edu.ntnu.idi.idatt.domain.entity.Player;
+import edu.ntnu.idi.idatt.utils.Validation;
 
-    /**
-     * SnakeAction class is a class that represents the action of a tile with a skip-turn effect.
-     * When a player lands on a tile with a skip-turn effect, the player will skip their next turn.
-     *
-     * @version 0.2
-     * @since 0.1
-     * @author Trang Duong
-     * @author Gilianne Reyes
-     */
-    public class SkipTurnAction implements TileAction {
-        public static final String actionType = "SkipTurnAction";
+/**
+ * SnakeAction class is a class that represents the action of a tile with a skip-turn effect.
+ * When a {@link Player} lands on a tile with a skip-turn effect,
+ * the player will skip their next turn.
+ *
+ * <br>
+ * <b>Note:</b> The player's skip-turn flag will be set to {@code true}.
+ *
+ * @version 0.2
+ * @since 0.1
+ * @author Trang Duong
+ * @author Gilianne Reyes
+ */
+public class SkipTurnAction implements TileAction {
+  public static final String actionType = "SkipTurnAction";
 
-        /**
-         * Constructs a SkipTurnAction instance.
-         */
-        public SkipTurnAction() {
-        }
+  /**
+   * Applies the skip-turn effect to the specified player.
+   *
+   * @param player is the player who landed on the skip-turn tile.
+   *
+   * @throws IllegalArgumentException if the player is {@code null}.
+   */
+  public void perform(Player player) {
+    Validation.validateNonNull(player, "Player");
+    player.setSkipTurn(true);
+  }
 
-        /**
-         * Sets the player to skip their next turn.
-         *
-         * @param player is the player that landed on the tile.
-         *
-         * @throws IllegalArgumentException if the player is null.
-         */
-        public void perform(Player player) {
-            Validation.validateNonNull(player, "Player");
-            player.setSkipTurn(true);
-        }
-
-        /**
-         * Retrieves the type of the action, which is "SkipTurnAction".
-         *
-         * @return the type of the action.
-         */
-        @Override
-        public String getActionType() {
-            return actionType;
-        }
-    }
+  /**
+   * Retrieves a string identifier describing the {@link SkipTurnAction}.
+   *
+   * @return a non-null string representing the action type.
+   */
+  @Override
+  public String getActionType() {
+    return actionType;
+  }
+}
